+++ synq/base_api_test.go
@@ -1,6 +1,7 @@
 package synq
 
 import (
+	"bytes"
 	"errors"
 	"io/ioutil"
 	"log"
@@ -113,12 +114,103 @@ func setupTestServer(type_ ...string) {
 	}
 }
 
-func setupTestApi(key string, type_ ...string) Api {
-	api := Api{}
-	api.Key = key
-	setupTestServer(type_...)
-	api.Url = testServer.URL
-	return api
+func TestParseAwsResp(t *testing.T) {
+	assert := assert.New(t)
+	var v interface{}
+	resp := http.Response{
+		StatusCode: 204,
+	}
+	err := errors.New("failure")
+	e := parseAwsResp(&resp, err, v)
+	assert.NotNil(e)
+	assert.Equal("failure", e.Error())
+
+	br := BadReader{}
+	resp = http.Response{
+		StatusCode: 412,
+		Body:       ioutil.NopCloser(br),
+	}
+	e = parseAwsResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("failed to read", e.Error())
+
+	err_msg := loadSample("upload.json")
+	resp = http.Response{
+		StatusCode: 412,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(err_msg)),
+	}
+	e = parseAwsResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("EOF", e.Error())
+
+	err_msg = loadSample("aws_err.xml")
+	resp = http.Response{
+		StatusCode: 412,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(err_msg)),
+	}
+	e = parseAwsResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("At least one of the pre-conditions you specified did not hold", e.Error())
+
+	resp = http.Response{
+		StatusCode: 204,
+	}
+	e = parseAwsResp(&resp, nil, v)
+	assert.Nil(e)
+}
+
+func TestParseSynqResp(t *testing.T) {
+	assert := assert.New(t)
+	var v interface{}
+	resp := http.Response{
+		StatusCode: 200,
+	}
+	err := errors.New("failure")
+	e := parseSynqResp(&resp, err, v)
+	assert.NotNil(e)
+	assert.Equal("failure", e.Error())
+	br := BadReader{}
+	resp = http.Response{
+		StatusCode: 200,
+		Body:       ioutil.NopCloser(br),
+	}
+	e = parseSynqResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("failed to read", e.Error())
+	err_msg := loadSample("aws_err.xml")
+	resp = http.Response{
+		StatusCode: 400,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(err_msg)),
+	}
+	e = parseSynqResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("could not parse : <Error>\n  <Code>PreconditionFailed</Code>\n  <Message>At least one of the pre-conditions you specified did not hold</Message>\n  <Condition>Bucket POST must be of the enclosure-type multipart/form-data</Condition>\n  <RequestId>634081169DAFE345</RequestId>\n  <HostId>80jHDkIWiVJd6ofogZSnvEfIxEUk35ULsvWPYFcH5f6VSUMPhCAevKwzLWN+Iw6gGTEvgogepSY=</HostId>\n</Error>\n", e.Error())
+	err_msg = []byte(INVALID_UUID)
+	resp = http.Response{
+		StatusCode: 400,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(err_msg)),
+	}
+	e = parseSynqResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("Invalid uuid. Example: '1c0e3ea4529011e6991554a050defa20'.", e.Error())
+	msg := []byte("<xml>")
+	resp = http.Response{
+		StatusCode: 200,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(msg)),
+	}
+	e = parseSynqResp(&resp, nil, v)
+	assert.NotNil(e)
+	assert.Equal("could not parse : <xml>", e.Error())
+	msg = loadSample("video.json")
+	var video Video
+	resp = http.Response{
+		StatusCode: 200,
+		Body:       ioutil.NopCloser(bytes.NewBuffer(msg)),
+	}
+	e = parseSynqResp(&resp, nil, &video)
+	assert.Nil(e)
+	assert.Equal(VIDEO_ID, video.Id)
+	assert.NotEmpty(video.Input)
 }
 
 func TestNew(t *testing.T) {
